Fix focus_bydirection when the focussed client in below another one
While having a partially obscured focussed client is a valid use case, having
having to hit two shortcut while navigating with keyboard is not.

The only valid senario when the current behavior is wanter is navigating across
multiple window. In "that" case, it might be a good idea to preserve the stack,
 but I don't think this commit will do more harm than it solve.